bt: split bt.c into smaller pieces
Moved from one large bt.c file into several files roughly
corresponding with a single phase of the sync process:
- scan.c
- connect.c
- uplink.c
- downlink.c
