+++ gateway/CMakeLists.txt
@@ -4,9 +4,14 @@ cmake_minimum_required(VERSION 3.20.0)
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(gateway)
 
+target_include_directories(app PRIVATE include)
+
 target_sources(app PRIVATE
   src/block.c
-  src/bt.c
+  src/bt/connect.c
+  src/bt/downlink.c
+  src/bt/scan.c
+  src/bt/uplink.c
   src/downlink.c
   src/uplink.c
   src/main.c
