+++ gateway/include/gateway/bt/scan.h
@@ -1,0 +1,7 @@
+/*
+ * Copyright (c) 2025 Golioth, Inc.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+void gateway_scan_start(void);
