--- gateway/src/bt.h
@@ -1,12 +1,0 @@
-/*
- * Copyright (c) 2025 Golioth, Inc.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-#ifndef __BT_APP_H__
-#define __BT_APP_H__
-
-int bt_app_start(void);
-
-#endif /* __BT_APP_H__ */
